tech(redis_config): extract all redis setup in a dedicated redis_config file
