+++ src/lib.rs
@@ -5,5 +5,6 @@ pub mod file;
 pub mod http;
 pub mod keyring;
 pub mod keyring_utils;
+pub mod redis_config;
 pub mod redis_utils;
 pub mod write_once_service;
