+++ tests/ensure_write_once.rs
@@ -3,8 +3,8 @@ extern crate ds_proxy;
 use actix_web::guard::Get;
 use actix_web::web::resource;
 use actix_web::HttpResponse;
-use ds_proxy::config::RedisConfig;
 use ds_proxy::http::middlewares::ensure_write_once;
+use ds_proxy::redis_config::RedisConfig;
 use std::thread;
 use url::Url;
 mod helpers;
